chore: moved ast to parser crate
related to code organization
