+++ crates/hir/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2024"
 [dependencies]
 color-eyre.workspace=true
 common = {path="../common"}
+slynx-parser = {path="../parser"}
