+++ crates/parser/src/ast/mod.rs
@@ -2,10 +2,10 @@ mod component;
 mod expression;
 mod types;
 
+use common::Span;
 pub use component::*;
 pub use expression::*;
 pub use types::*;
-
 #[derive(Default, Debug, Clone, Copy)]
 pub enum VisibilityModifier {
     ///Property visible to everyone
@@ -17,42 +17,6 @@ pub enum VisibilityModifier {
     ChildrenPublic,
     ///Property visible only for the parents. Only usable on Components.
     ParentPublic,
-}
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-///The representation of the bounds of something on the code.
-pub struct Span {
-    pub start: usize,
-    pub end: usize,
-}
-
-impl Span {
-    ///Merges this span with the given `target`. The returned span will have the initial position of this one, and the final position of the given `target`
-    pub fn merge_with(mut self, target: Self) -> Self {
-        self.end = target.end;
-        self
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-#[repr(C)]
-///Some operator on the code. Something like, +, - , *, /, &, &&, etc
-pub enum Operator {
-    Add,
-    Sub,
-    Star,
-    Slash,
-    Equals,
-    GreaterThan,
-    GreaterThanOrEqual,
-    LessThan,
-    LessThanOrEqual,
-    LogicAnd,
-    LogicOr,
-    And,
-    Or,
-    RightShift,
-    LeftShift,
-    Xor,
 }
 
 #[derive(Debug)]
@@ -127,21 +91,6 @@ impl ASTExpression {
         matches!(
             self.kind,
             ASTExpressionKind::Identifier(_) | ASTExpressionKind::FieldAccess { .. },
-        )
-    }
-}
-
-impl Operator {
-    pub fn is_logical(&self) -> bool {
-        matches!(
-            self,
-            Self::LogicAnd
-                | Self::LogicOr
-                | Self::Equals
-                | Self::GreaterThan
-                | Self::GreaterThanOrEqual
-                | Self::LessThan
-                | Self::LessThanOrEqual
         )
     }
 }
