+++ crates/parser/src/expr.rs
@@ -1,9 +1,10 @@
+use crate::{
+    ASTExpression, ASTExpressionKind, ComponentExpression, ComponentMemberValue, GenericIdentifier,
+    NamedExpr,
+};
 use crate::{Parser, error::ParseError};
 use color_eyre::eyre::Result;
-use common::ast::{
-    ASTExpression, ASTExpressionKind, ComponentExpression, ComponentMemberValue, GenericIdentifier,
-    NamedExpr, Operator, Span,
-};
+use common::{Operator, Span};
 use slynx_lexer::tokens::{Token, TokenKind};
 
 impl Parser {
